Refactoring. Correction de l'accès aux adhésions pour le ROLE_FINANCE_MANAGER
