+++ app/Resources/views/admin/index.html.twig
@@ -19,7 +19,7 @@
             <a href="{{ path("admin_formations") }}" class="waves-effect waves-light btn"><i class="material-icons left">tune</i>Gérer les formations</a>
 
             <h6>Finances</h6>
-            <a href="{{ path("admin_registrations") }}" class="waves-effect waves-light btn indigo"><i class="material-icons left">attach_money</i>Gérer les Adhésions / Réadhésions</a>
+            <a href="{{ path("registrations") }}" class="waves-effect waves-light btn indigo"><i class="material-icons left">attach_money</i>Gérer les Adhésions / Réadhésions</a>
             <a href="{{ path("helloasso_payments") }}" class="waves-effect waves-light btn indigo"><i class="material-icons left">settings_ethernet</i>Paiements helloasso</a>
 
             <h6>Créneaux</h6>
