+++ src/AppBundle/Controller/AdminController.php
@@ -232,387 +232,6 @@ class AdminController extends Controller
     }
 
     /**
-     * Registrations list
-     *
-     * @Route("/registrations", name="admin_registrations")
-     * @Method({"POST","GET"})
-     * @Security("has_role('ROLE_FINANCE_MANAGER')")
-     */
-    public function registrationsAction(Request $request)
-    {
-        $session = new Session();
-
-        $qfrom = $request->query->get('from');
-        if (!$qfrom) {
-            $monday = strtotime('last monday', strtotime('tomorrow'));
-            $from = new DateTime();
-            $from->setTimestamp($monday);
-        }else{
-            $from = date_create_from_format('Y-m-d', $qfrom );
-            if (!$from || $from->format('Y-m-d') != $qfrom) {
-                $session->getFlashBag()->add('warning','la date "'.$qfrom.'"" n\'est pas au bon format (Y-m-d)');
-                $monday = strtotime('last monday', strtotime('tomorrow'));
-                $from = new DateTime();
-                $from->setTimestamp($monday);
-            }
-        }
-        $from = $from->setTime('0','0','0');
-
-        $qto = $request->query->get('to');
-        if ($qto) {
-            $to = date_create_from_format('Y-m-d', $qto );
-            if (!$to || $to->format('Y-m-d') != $qto) {
-                $session->getFlashBag()->add('warning','la date "'.$qto.'"" n\'est pas au bon format (Y-m-d)');
-                $to = null;
-            }else{
-                $to = $to->setTime('0','0','0');
-            }
-        }else{
-            $to = null;
-        }
-
-
-        $em = $this->getDoctrine()->getManager();
-        if (!($page = $request->get('page')))
-            $page = 1;
-        $limit = 25;
-        $qb = $em->createQueryBuilder()->from('AppBundle\Entity\AbstractRegistration', 'r')
-            ->select('count(r.id)')
-            ->where('r.date >= :from')
-            ->setParameter('from', $from);
-        if ($to){
-            $qb = $qb->andwhere('r.date <= :to')->setParameter('to', $to);
-        }
-
-        $max = $qb->getQuery()
-            ->getSingleScalarResult();
-        $nb_of_pages = intval($max / $limit);
-        $nb_of_pages += (($max % $limit) > 0) ? 1 : 0;
-        $repository = $em->getRepository('AppBundle:AbstractRegistration');
-        $queryb = $repository->createQueryBuilder('r')
-            ->where('r.date >= :from')
-            ->setParameter('from', $from);
-        if ($to){
-            $queryb = $queryb->andwhere('r.date <= :to')->setParameter('to', $to);
-        }
-        $queryb = $queryb->orderBy('r.date', 'DESC')
-            ->setFirstResult(($page - 1) * $limit)
-            ->setMaxResults($limit);
-
-        $registrations = $queryb->getQuery()->getResult();
-        $delete_forms = array();
-
-        $table_name = $em->getClassMetadata('AppBundle:AbstractRegistration')->getTableName();
-        $connection = $em->getConnection();
-        $statement = $connection->prepare("SELECT date, SUM(sum_1) as sum_1,SUM(sum_2) as sum_2,SUM(sum_3) as sum_3,SUM(sum_4) as sum_4,SUM(sum_5) as sum_5,SUM(sum_6) as sum_6,SUM(grand_total) as grand_total FROM
-(SELECT date_format(date,\"%Y-%m-%d\") as date,
-SUM(IF(mode='1',amount,0)) as sum_1,
-SUM(IF(mode='2',amount,0)) as sum_2,
-SUM(IF(mode='3',amount,0)) as sum_3,
-SUM(IF(mode='4',amount,0)) as sum_4,
-SUM(IF(mode='5',amount,0)) as sum_5,
-SUM(IF(mode='6',amount,0)) as sum_6,
-SUM(amount) as grand_total
-FROM ".$table_name."
-WHERE date >= :from ".(($to) ? "AND date <= :to" : "")."
-GROUP BY date) as sum GROUP BY date ORDER BY date DESC;");
-        $statement->bindValue('from', $from->format('Y-m-d'));
-        if ($to){
-            $statement->bindValue('to', $to->format('Y-m-d'));
-        }
-        $statement->execute();
-        $results = $statement->fetchAll();
-
-        $totaux = array();
-        foreach ($results as $result){
-            $totaux[$result['date']] = $result;
-        }
-
-        $connection = $em->getConnection();
-        $statement = $connection->prepare("SELECT
-SUM(IF(mode='1',amount,0)) as sum_1,
-SUM(IF(mode='2',amount,0)) as sum_2,
-SUM(IF(mode='3',amount,0)) as sum_3,
-SUM(IF(mode='4',amount,0)) as sum_4,
-SUM(IF(mode='5',amount,0)) as sum_5,
-SUM(IF(mode='6',amount,0)) as sum_6,
-SUM(amount) as grand_total
-FROM ".$table_name."
-WHERE date >= :from ".(($to) ? "AND date <= :to" : "").";");
-        $statement->bindValue('from', $from->format('Y-m-d'));
-        if ($to){
-            $statement->bindValue('to', $to->format('Y-m-d'));
-        }
-        $statement->execute();
-        $grand_total = $statement->fetch();
-
-
-        $re = '/1_([0-9]+)$/m';
-        foreach ($registrations as $registration) {
-            if ($registration->getType() == AbstractRegistration::TYPE_MEMBER){
-                $matches = array();
-                if (preg_match($re, $registration->getId(), $matches)) {
-                    $delete_forms[$registration->getId()] = $this->getRegistrationDeleteForm($matches[1])->createView();
-                }
-            }
-        }
-
-        return $this->render('admin/registrations/list.html.twig',
-            array(
-                'R' => new Registration(),
-                'registrations' => $registrations,
-                'grand_total' => $grand_total,
-                'totaux' => $totaux,
-                'delete_forms' => $delete_forms,
-                'page' => $page,
-                'from' => $from,
-                'to' => $to,
-                'nb_of_pages' => $nb_of_pages));
-    }
-
-    /**
-     * edit registration
-     *
-     * @Route("/registration/{id}/edit", name="admin_registration_edit")
-     * @Method({"GET","POST"})
-     * @Security("has_role('ROLE_FINANCE_MANAGER')")
-     */
-    public function editRegistrationAction(Request $request, Registration $registration)
-    {
-        $session = new Session();
-        if ($registration->getId() && ($request->attributes->get('id') == $registration->getId())){
-            $edit_form = $this->createForm(RegistrationType::class, $registration);
-            $edit_form->handleRequest($request);
-            if ($edit_form->isSubmitted() && $edit_form->isValid()) {
-                $em = $this->getDoctrine()->getManager();
-                $em->persist($registration);
-                $em->flush();
-                $session->getFlashBag()->add('success', 'La ligne a bien été éditée !');
-                return $this->redirectToRoute("admin_registrations");
-            }
-        }else{
-            $session->getFlashBag()->add('error', 'l\'entrée #'.$request->attributes->get('id').' n\'a pas été trouvée');
-            return $this->redirectToRoute("admin_registrations");
-        }
-
-        return $this->render('admin/registrations/edit.html.twig', array('edit_form' => $edit_form->createView(),'registration' => $registration));
-    }
-
-    /**
-     * remove registration
-     *
-     * @Route("/registration/{id}/remove", name="admin_registration_remove")
-     * @Method({"DELETE"})
-     * @Security("has_role('ROLE_SUPER_ADMIN')")
-     */
-    public function removeRegistrationAction(Request $request, Registration $registration)
-    {
-        $session = new Session();
-        $form = $this->getRegistrationDeleteForm($registration->getId());
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            if ($registration->getUser() && count($registration->getUser()->getRegistrations()) === 1 && $registration === $registration->getUser()->getLastRegistration()) {
-                $session->getFlashBag()->add('error', 'C\'est la seule adhésion de cette adhérent, corrigez là plutôt que de la supprimer');
-            } else {
-                $em = $this->getDoctrine()->getManager();
-                if ($registration->getUser()) {
-                    $registration->getUser()->removeRegistration($registration);
-                    $em->persist($registration->getUser());
-                }
-                if ($registration->getRegistrar()) {
-                    $registration->getRegistrar()->removeRecordedRegistration($registration);
-                    $em->persist($registration->getRegistrar());
-                }
-                $em->remove($registration);
-                $em->flush();
-                $session->getFlashBag()->add('success', 'L\'adhésion a bien été supprimée !');
-            }
-        }
-        return $this->redirectToRoute('admin_registrations');
-    }
-
-    /**
-     * @param integer $registration_id
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    protected function getRegistrationDeleteForm(int $registration_id)
-    {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('admin_registration_remove', array('id' => $registration_id)))
-            ->setMethod('DELETE')
-            ->getForm();
-    }
-
-    /**
-     * Helloasso payments list
-     *
-     * @Route("/helloassoPayments", name="helloasso_payments")
-     * @Method("GET")
-     * @Security("has_role('ROLE_FINANCE_MANAGER')")
-     */
-    public function helloassoPaymentsAction(Request $request)
-    {
-        if (!($page = $request->get('page')))
-            $page = 1;
-        $limit = 50;
-        $max = $this->getDoctrine()->getManager()->createQueryBuilder()->from('AppBundle\Entity\HelloassoPayment', 'n')
-            ->select('count(n.id)')
-            ->getQuery()
-            ->getSingleScalarResult();
-        $nb_of_pages = intval($max / $limit);
-        if ($max > 0)
-            $nb_of_pages += (($max % $limit) > 0) ? 1 : 0;
-        $payments = $this->getDoctrine()->getManager()
-            ->getRepository('AppBundle:HelloassoPayment')
-            ->findBy(array(), array('created_at' => 'DESC', 'date' => 'DESC'), $limit, ($page - 1) * $limit);
-        $delete_forms = array();
-        foreach ($payments as $payment) {
-            $delete_forms[$payment->getId()] = $this->getPaymentDeleteForm($payment)->createView();
-        }
-
-        //todo: save this somewhere ?
-        $campaigns_json = $this->container->get('AppBundle\Helper\Helloasso')->get('campaigns');
-        $campaigns = array();
-        foreach ($campaigns_json->resources as $c) {
-            $campaigns[intval($c->id)] = $c;
-        }
-
-        return $this->render(
-            'admin/helloasso/payments.html.twig',
-            array('payments' => $payments,
-                'campaigns' => $campaigns,
-                'delete_forms' => $delete_forms,
-                'page' => $page,
-                'nb_of_pages' => $nb_of_pages));
-    }
-
-    /**
-     * Helloasso browser
-     *
-     * @Route("/helloassoBrowser", name="helloasso_browser")
-     * @Method("GET")
-     * @Security("has_role('ROLE_FINANCE_MANAGER')")
-     */
-    public function helloassoBrowserAction(Request $request)
-    {
-        if (!($page = $request->get('page')))
-            $page = 1;
-
-        if (!($campaignId = $request->get('campaign'))) {
-            $campaigns_json = $this->container->get('AppBundle\Helper\Helloasso')->get('campaigns');
-            $campaigns = $campaigns_json->resources;
-            return $this->render(
-                'admin/helloasso/browser.html.twig',
-                array('campaigns' => $campaigns));
-        } else {
-            $campaignId = str_pad($campaignId, 12, '0', STR_PAD_LEFT);
-            $campaign_json = $this->container->get('AppBundle\Helper\Helloasso')->get('campaigns/' . $campaignId);
-            $payments_json = $this->container->get('AppBundle\Helper\Helloasso')->get('campaigns/' . $campaignId . '/payments', array('page' => $page));
-            $page = $payments_json->pagination->page;
-            $nb_of_pages = $payments_json->pagination->max_page;
-            $results_per_page = $payments_json->pagination->results_per_page;
-            return $this->render(
-                'admin/helloasso/browser.html.twig',
-                array('payments' => $payments_json->resources,
-                    'page' => $page,
-                    'campaign' => $campaign_json,
-                    'nb_of_pages' => $nb_of_pages));
-        }
-
-    }
-
-    /**
-     * Helloasso manual paiement add
-     *
-     * @Route("/helloassoManualPaimentAdd/", name="helloasso_manual_paiement_add")
-     * @Method("POST")
-     * @Security("has_role('ROLE_FINANCE_MANAGER')")
-     */
-    public function helloassoManualPaimentAddAction(Request $request)
-    {
-        $session = new Session();
-        if (!($paiementId = $request->get('paiementId'))) {
-            $session->getFlashBag()->add('error', 'missing paiment id');
-            return $this->redirectToRoute('helloasso_browser');
-        } else {
-            $payment_json = $this->container->get('AppBundle\Helper\Helloasso')->get('payments/' . $paiementId);
-
-            $em = $this->getDoctrine()->getManager();
-            $exist = $em->getRepository('AppBundle:HelloassoPayment')->findOneBy(array('paymentId' => $payment_json->id));
-
-            if ($exist) {
-                $session->getFlashBag()->add('error', 'Ce paiement est déjà enregistré');
-                return $this->redirectToRoute('helloasso_browser', array('campaign' => $exist->getCampaignId()));
-            }
-
-            $payments = array();
-            $action_json = null;
-            $dispatcher = $this->get('event_dispatcher');
-            foreach ($payment_json->actions as $action) {
-                $action_json = $this->container->get('AppBundle\Helper\Helloasso')->get('actions/' . $action->id);
-                $payment = $em->getRepository('AppBundle:HelloassoPayment')->findOneBy(array('paymentId' => $payment_json->id));
-                if ($payment) { //payment already exist (created from a previous actions in THIS loop)
-                    $amount = $action_json->amount;
-                    $amount = str_replace(',', '.', $amount);
-                    $payment->setAmount($payment->getAmount() + $amount);
-                } else {
-                    $payment = new HelloassoPayment();
-                    $payment->fromActionObj($action_json);
-                }
-                $em->persist($payment);
-                $em->flush();
-                $payments[$payment->getId()] = $payment;
-            }
-            foreach ($payments as $payment) {
-                $dispatcher->dispatch(
-                    HelloassoEvent::PAYMENT_AFTER_SAVE,
-                    new HelloassoEvent($payment)
-                );
-            }
-
-            $session->getFlashBag()->add('success', 'Ce paiement a bien été enregistré');
-            return $this->redirectToRoute('helloasso_browser', array('campaign' => $action_json->id_campaign));
-        }
-    }
-
-    /**
-     * remove payment
-     *
-     * @Route("/helloasso/{id}", name="helloasso_payment_remove")
-     * @Method({"DELETE"})
-     * @Security("has_role('ROLE_SUPER_ADMIN')")
-     */
-    public function removePaymentAction(Request $request, HelloassoPayment $payment)
-    {
-        $session = new Session();
-        $form = $this->getPaymentDeleteForm($payment);
-        $form->handleRequest($request);
-        if ($form->isSubmitted() && $form->isValid()) {
-            $em = $this->getDoctrine()->getManager();
-            if ($payment->getRegistration()) {
-                $session->getFlashBag()->add('error', 'ce paiement est lié à une adhésion');
-                return $this->redirectToRoute('helloasso_payments');
-            }
-            $em->remove($payment);
-            $em->flush();
-            $session->getFlashBag()->add('success', 'Le paiement a bien été supprimé !');
-        }
-        return $this->redirectToRoute('helloasso_payments');
-    }
-
-    /**
-     * @param HelloassoPayment $payment
-     * @return \Symfony\Component\Form\FormInterface
-     */
-    protected function getPaymentDeleteForm(HelloassoPayment $payment)
-    {
-        return $this->createFormBuilder()
-            ->setAction($this->generateUrl('helloasso_payment_remove', array('id' => $payment->getId())))
-            ->setMethod('DELETE')
-            ->getForm();
-    }
-
-    /**
      * Widget generator
      *
      * @Route("/widget", name="widget_generator")
